Extract Token Store cache into a separate class
